requirements.txt
gradio==4.40.0
imageio-ffmpeg==0.5.1
torch==2.2.0
git+https://github.com/huggingface/diffusers.git@878f609aa5ce4a78fea0f048726889debde1d7e8#egg=diffusers
bitsandbytes
transformers==4.42.0
spaces==0.29.2
moviepy==1.0.3
openai==1.38.0
ollama
accelerate==0.33.0
sentencepiece==0.2.0
Pillow==9.5.0